+++ Tests.Common/Rounding/TestIntegerRoundingDown.cs
@@ -7,79 +7,79 @@ namespace Tests.Common.Rounding
 	[TestFixture ()]
 	public class TestIntegerRoundingDown
 	{
-		private const int TEST_POS_RESULT_NUMBER = 1000;
+		private const Int32 TEST_POS_RESULT_NUMBER = 1000;
 
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_01_Cents()
 		{
 			decimal testInputs = 1000.01m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_20_Cents()
 		{
 			decimal testInputs = 1000.20m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_50_Cents()
 		{
 			decimal testInputs = 1000.50m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_60_Cents()
 		{
 			decimal testInputs = 1000.60m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_99_Cents()
 		{
 			decimal testInputs = 1000.99m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 
-		private const int TEST_NEG_RESULT_NUMBER = -1000;
+		private const Int32 TEST_NEG_RESULT_NUMBER = -1000;
 
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_01_Cents()
 		{
 			decimal testInputs = -1000.01m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_20_Cents()
 		{
 			decimal testInputs = -1000.20m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_50_Cents()
 		{
 			decimal testInputs = -1000.50m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_60_Cents()
 		{
 			decimal testInputs = -1000.60m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_99_Cents()
 		{
 			decimal testInputs = -1000.99m;
-			long testResult = IntRounding.RoundDown(testInputs);
+			Int32 testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 	}
